+++ example/integration_test/app_test.dart
@@ -1,5 +1,6 @@
 import 'package:integration_test/integration_test.dart';
 
+import 'accessibility_scenarios.dart';
 import 'app_smoke_scenarios.dart';
 import 'context_menu_scenarios.dart';
 import 'drag_scenarios.dart';
@@ -19,6 +20,7 @@ import 'snapping_scenarios.dart';
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
+  accessibilityScenarios();
   appSmokeScenarios();
   contextMenuScenarios();
   dragScenarios();